+++ src/main/java/com/minecrafttas/tasmodog/virtual/structs/KeyEvent.java
@@ -1,4 +1,4 @@
-package com.minecrafttas.tasmodog.structs;
+package com.minecrafttas.tasmodog.virtual.structs;
 
 import java.io.Serializable;
